+++ targetCity.php
@@ -1,0 +1,7 @@
+<html>
+    <div>
+        <p>Pour ajouter une ville, merci de cliquer <a href="addCity.php" target="_blank">Ici</a></p>
+    </div>
+</html>
+
+
